move the version number to it's own module in the package
This will make the version number available in the package itself, not
only in the setup.py.

The version numbering is based on https://github.com/silx-kit/silx/blob/master/src/silx/_version.py
To best make use of it, new releases should be done one their own branch
from now on, not on master. On the master branch, we increment MAJOR,
MINOR and MICRO as needed, after each release cut. On the release
branch, we can increment RELEV and SERIAL (unlikely).